+++ src/lmi/waypoint/db/WpSegmentRecord.java
@@ -4,10 +4,10 @@ public final class WpSegmentRecord {
   public final long id;
   public final long edgeId;
   public final long graphId;
-  public final long cutId;
+  public final int cutId;
   public final int step;
 
-  public WpSegmentRecord(long id, long edgeId, long graphId, long cutId, int step) {
+  public WpSegmentRecord(long id, long edgeId, long graphId, int cutId, int step) {
     this.id = id;
     this.edgeId = edgeId;
     this.graphId = graphId;
