+++ src/lmi/waypoint/model/LoadPointsByCutResult.java
@@ -6,10 +6,10 @@ import lmi.waypoint.persistence.WaypointResult;
 
 public final class LoadPointsByCutResult implements WaypointResult {
   public final long graphId;
-  public final long cutId;
+  public final int cutId;
   public final Array<WpPoint> points;
 
-  public LoadPointsByCutResult(long graphId, long cutId, Array<WpPoint> points) {
+  public LoadPointsByCutResult(long graphId, int cutId, Array<WpPoint> points) {
     this.graphId = graphId;
     this.cutId = cutId;
     this.points = points;
