+++ src/lmi/waypoint/model/LoadSegmentsByCutResult.java
@@ -6,10 +6,10 @@ import lmi.waypoint.persistence.WaypointResult;
 
 public final class LoadSegmentsByCutResult implements WaypointResult {
   public final long graphId;
-  public final long cutId;
+  public final int cutId;
   public final Array<WpSegment> segments;
 
-  public LoadSegmentsByCutResult(long graphId, long cutId, Array<WpSegment> segments) {
+  public LoadSegmentsByCutResult(long graphId, int cutId, Array<WpSegment> segments) {
     this.graphId = graphId;
     this.cutId = cutId;
     this.segments = segments;
